refactor: index added for services
