+++ src/services/index.js
@@ -1,0 +1,5 @@
+export {API_URL} from "./apiDetails";
+export {fetchAllProducts} from "./products";
+export * from "./address";
+export * from "./cart";
+export * from "./wishlist";
